https://mimarlik-sitesi.netlify.app/
This project is an architecture-themed website created to improve my web development skills. Built with HTML and CSS, it features a responsive banner, navigation menu, and buttons with hover effects. The 'Design Your Home' section provides information and interactive options for users. The design was created to strengthen my front-end skills and gain practical experience.